book_store/requirements.txt
Django==2.2.8
django-environ==0.8.1
psycopg2-binary==2.9.3
pylint==2.12.2
pytz==2021.3
sqlparse==0.4.2

djangorestframework==3.11.0
djangorestframework-simplejwt==4.8.0